Rename CGDisplayList and friends to their real names https://bugs.webkit.org/show_bug.cgi?id=261938 rdar://115880900
Reviewed by Wenson Hsieh.

Rename and upstream various dynamic-content-scaling-related classes.
Also, remove a very noisy bit of logging, and a setting that is no longer needed.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/graphics/ImageBuffer.h:
* Source/WebCore/platform/graphics/cocoa/WebCoreCALayerExtras.h:
* Source/WebCore/platform/graphics/cocoa/WebCoreCALayerExtras.mm:
(-[CALayer _web_clearContents]):
(-[CALayer _web_clearDynamicContentScalingDisplayListIfNeeded]):
(-[CALayer _web_clearCGDisplayListIfNeeded]): Deleted.
* Source/WebCore/platform/graphics/re/DynamicContentScalingTypes.h: Copied from Source/WebCore/platform/graphics/cocoa/WebCoreCALayerExtras.h.
* Source/WebKit/Shared/RemoteLayerTree/DynamicContentScalingDisplayList.cpp: Renamed from Source/WebKit/Shared/RemoteLayerTree/CGDisplayList.cpp.
(WebKit::DynamicContentScalingDisplayList::encode):
(WebKit::DynamicContentScalingDisplayList::decode):
* Source/WebKit/Shared/RemoteLayerTree/DynamicContentScalingDisplayList.h: Renamed from Source/WebKit/Shared/RemoteLayerTree/CGDisplayList.h.
(WebKit::DynamicContentScalingDisplayList::DynamicContentScalingDisplayList):
(WebKit::DynamicContentScalingDisplayList::buffer const):
(WebKit::DynamicContentScalingDisplayList::takeSurfaces):
* Source/WebKit/Shared/RemoteLayerTree/DynamicContentScalingImageBufferBackend.h: Renamed from Source/WebKit/Shared/RemoteLayerTree/CGDisplayListImageBufferBackend.h.
* Source/WebKit/Shared/RemoteLayerTree/DynamicContentScalingImageBufferBackend.mm: Renamed from Source/WebKit/Shared/RemoteLayerTree/CGDisplayListImageBufferBackend.mm.
(WTF::CFTypeTrait<CAMachPortRef>::typeID):
(WebKit::makeContextOptions):
(WebKit::GraphicsContextDynamicContentScaling::GraphicsContextDynamicContentScaling):
(WebKit::DynamicContentScalingImageBufferBackend::calculateMemoryCost):
(WebKit::DynamicContentScalingImageBufferBackend::create):
(WebKit::DynamicContentScalingImageBufferBackend::DynamicContentScalingImageBufferBackend):
(WebKit::m_renderingMode):
(WebKit::DynamicContentScalingImageBufferBackend::createBackendHandle const):
(WebKit::DynamicContentScalingImageBufferBackend::context):
(WebKit::DynamicContentScalingImageBufferBackend::backendSize const):
(WebKit::DynamicContentScalingImageBufferBackend::bytesPerRow const):
(WebKit::DynamicContentScalingImageBufferBackend::releaseGraphicsContext):
(WebKit::DynamicContentScalingImageBufferBackend::copyNativeImage):
(WebKit::DynamicContentScalingImageBufferBackend::createNativeImageReference):
(WebKit::DynamicContentScalingImageBufferBackend::getPixelBuffer):
(WebKit::DynamicContentScalingImageBufferBackend::putPixelBuffer):
(WebKit::DynamicContentScalingImageBufferBackend::debugDescription const):
(WebKit::DynamicContentScalingAcceleratedImageBufferBackend::create):
(WebKit::DynamicContentScalingAcceleratedImageBufferBackend::DynamicContentScalingAcceleratedImageBufferBackend):
* Source/WebKit/Shared/RemoteLayerTree/RemoteLayerBackingStore.h:
* Source/WebKit/Shared/RemoteLayerTree/RemoteLayerBackingStore.mm:
(WebKit::RemoteLayerBackingStore::clearBackingStore):
(WebKit::hasValue):
(WebKit::RemoteLayerBackingStore::encode const):
(WebKit::RemoteLayerBackingStoreProperties::decode):
(WebKit::RemoteLayerBackingStore::swapToValidFrontBuffer):
(WebKit::RemoteLayerBackingStore::supportsPartialRepaint const):
(WebKit::RemoteLayerBackingStore::ensureFrontBuffer):
(WebKit::RemoteLayerBackingStore::paintContents):
(WebKit::RemoteLayerBackingStore::drawInContext):
(WebKit::RemoteLayerBackingStoreProperties::layerContentsBufferFromBackendHandle):
(WebKit::RemoteLayerBackingStoreProperties::applyBackingStoreToLayer):
* Source/WebKit/Shared/RemoteLayerTree/RemoteLayerTreePropertyApplier.mm:
(WebKit::RemoteLayerTreePropertyApplier::applyPropertiesToLayer):
* Source/WebKit/SourcesCocoa.txt:
* Source/WebKit/UIProcess/RemoteLayerTree/RemoteLayerTreeHost.h:
* Source/WebKit/UIProcess/RemoteLayerTree/RemoteLayerTreeHost.mm:
(WebKit::RemoteLayerTreeHost::replayDynamicContentScalingDisplayListsIntoBackingStore const):
(WebKit::RemoteLayerTreeHost::replayCGDisplayListsIntoBackingStore const): Deleted.
* Source/WebKit/UIProcess/RemoteLayerTree/cocoa/RemoteLayerTreeLayers.h:
* Source/WebKit/UIProcess/RemoteLayerTree/cocoa/RemoteLayerTreeLayers.mm:
(-[WKCompositingLayer _setWKContents:withDisplayList:replayForTesting:]):
(-[WKCompositingLayer drawInContext:]):
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Source/WebKit/WebProcess/GPU/graphics/ImageBufferBackendHandle.h:
* Source/WebKit/WebProcess/WebPage/RemoteLayerTree/PlatformCALayerRemote.h:
* Source/WebKit/WebProcess/WebPage/RemoteLayerTree/PlatformCALayerRemote.mm:
(WebKit::PlatformCALayerRemote::shouldIncludeDisplayListInBackingStore const):
(WebKit::PlatformCALayerRemote::updateBackingStore):
* Source/WebKit/WebProcess/WebPage/RemoteLayerTree/RemoteLayerTreeContext.h:
(WebKit::RemoteLayerTreeContext::useDynamicContentScalingDisplayListsForDOMRendering const):
(WebKit::RemoteLayerTreeContext::setUseDynamicContentScalingDisplayListsForDOMRendering):
(WebKit::RemoteLayerTreeContext::useCGDisplayListsForDOMRendering const): Deleted.
(WebKit::RemoteLayerTreeContext::setUseCGDisplayListsForDOMRendering): Deleted.
(WebKit::RemoteLayerTreeContext::useCGDisplayListImageCache const): Deleted.
(WebKit::RemoteLayerTreeContext::setUseCGDisplayListImageCache): Deleted.
* Source/WebKit/WebProcess/WebPage/RemoteLayerTree/RemoteLayerTreeDrawingArea.mm:
(WebKit::RemoteLayerTreeDrawingArea::updatePreferences):

Canonical link: https://commits.webkit.org/268484@main
