+++ Source/WebKit/SourcesCocoa.txt
@@ -216,8 +216,8 @@ Shared/mac/MediaFormatReader/MediaSampleByteRange.cpp
 Shared/mac/MediaFormatReader/MediaSampleCursor.cpp
 Shared/mac/MediaFormatReader/MediaTrackReader.cpp
 
-Shared/RemoteLayerTree/CGDisplayList.cpp
-Shared/RemoteLayerTree/CGDisplayListImageBufferBackend.mm
+Shared/RemoteLayerTree/DynamicContentScalingDisplayList.cpp
+Shared/RemoteLayerTree/DynamicContentScalingImageBufferBackend.mm
 Shared/RemoteLayerTree/RemoteLayerBackingStore.mm
 Shared/RemoteLayerTree/RemoteLayerBackingStoreCollection.mm
 Shared/RemoteLayerTree/RemoteLayerWithRemoteRenderingBackingStoreCollection.mm
