+++ Source/WebKit/UIProcess/RemoteLayerTree/cocoa/RemoteLayerTreeLayers.mm
@@ -28,19 +28,20 @@
 
 #if PLATFORM(COCOA)
 
-#import "CGDisplayList.h"
+#import "DynamicContentScalingDisplayList.h"
 #import "Logging.h"
 #import "RemoteLayerTreeNode.h"
+#import <WebCore/DynamicContentScalingTypes.h>
 #import <pal/spi/cocoa/QuartzCoreSPI.h>
 #import <wtf/MachSendRight.h>
 #import <wtf/cocoa/TypeCastsCocoa.h>
 
-#if ENABLE(CG_DISPLAY_LIST_BACKED_IMAGE_BUFFER)
-#import <WebKitAdditions/CGDisplayListImageBufferAdditions.h>
+#if ENABLE(RE_DYNAMIC_CONTENT_SCALING)
+#import <CoreRE/RECGCommandsContext.h>
 #endif
 
 @implementation WKCompositingLayer {
-#if ENABLE(CG_DISPLAY_LIST_BACKED_IMAGE_BUFFER)
+#if ENABLE(RE_DYNAMIC_CONTENT_SCALING)
     RetainPtr<CFDataRef> _displayListDataForTesting;
 #endif
 }
@@ -50,9 +51,9 @@
     return WebKit::RemoteLayerTreeNode::appendLayerDescription(super.description, self);
 }
 
-#if ENABLE(CG_DISPLAY_LIST_BACKED_IMAGE_BUFFER)
+#if ENABLE(RE_DYNAMIC_CONTENT_SCALING)
 
-- (void)_setWKContents:(id)contents withDisplayList:(WebKit::CGDisplayList&&)displayList replayForTesting:(BOOL)replay
+- (void)_setWKContents:(id)contents withDisplayList:(WebKit::DynamicContentScalingDisplayList&&)displayList replayForTesting:(BOOL)replay
 {
     auto data = displayList.buffer()->createCFData();
 
@@ -72,8 +73,8 @@
         [ports addObject:static_cast<id>(portWrapper.get())];
     }
 
-    [self setValue:bridge_cast(data.get()) forKeyPath:WKCGDisplayListContentsKey];
-    [self setValue:ports.get() forKeyPath:WKCGDisplayListPortsKey];
+    [self setValue:bridge_cast(data.get()) forKeyPath:WKDynamicContentScalingContentsKey];
+    [self setValue:ports.get() forKeyPath:WKDynamicContentScalingContentsKey];
     [self setNeedsDisplay];
 }
 
@@ -83,10 +84,10 @@
         return;
     CGContextScaleCTM(context, 1, -1);
     CGContextTranslateCTM(context, 0, -self.bounds.size.height);
-    WKCGContextDrawCGCommandsEncodedData(context, _displayListDataForTesting.get(), nullptr);
+    RECGContextDrawCGCommandsEncodedData(context, _displayListDataForTesting.get(), nullptr);
 }
 
-#endif // ENABLE(CG_DISPLAY_LIST_BACKED_IMAGE_BUFFER)
+#endif // ENABLE(RE_DYNAMIC_CONTENT_SCALING)
 
 @end
